PP-4779: Move all gateway merchant id tests to its own test class
Additionally, did some refactoring on GatewayAccountResourceTestBase by making
methods `static` so the new GatewayMerchantIdITest class can re-use methods
without extending GatewayAccountResourceTestBase. This follows the "favor
composition over inheritance" principle.
